google-ads-loader/requirements.txt
google-ads>=26.0.0
anthropic>=0.40.0
requests>=2.32.0
python-dotenv>=1.0.0
pyyaml>=6.0.0
aiohttp>=3.9.0
streamlit>=1.35.0
